# Replace buildifier with ruff fmt