--- a/crates/pcb/Cargo.toml
+++ b/crates/pcb/Cargo.toml
@@ -33,7 +33,7 @@ pcb-kicad = { workspace = true }
 open = { workspace = true }
 inquire = { workspace = true }
 pcb-ui = { workspace = true }
-pcb-buildifier = { workspace = true }
+pcb-fmt = { workspace = true }
 pathdiff = { workspace = true }
 starlark = { workspace = true }
 starlark_syntax = { workspace = true }