--- a/crates/pcb/src/fmt.rs
+++ b/crates/pcb/src/fmt.rs
@@ -1,7 +1,7 @@
-use anyhow::{Context, Result};
+use anyhow::Result;
 use clap::Args;
 use log::debug;
-use pcb_buildifier::Buildifier;
+use pcb_fmt::RuffFormatter;
 use pcb_ui::prelude::*;
 use pcb_zen::file_extensions;
 use std::collections::HashSet;
@@ -9,10 +9,10 @@ use std::fs;
 use std::path::{Path, PathBuf};
 
 #[derive(Args, Debug, Default, Clone)]
-#[command(about = "Format .zen and .star files using buildifier")]
+#[command(about = "Format .zen files")]
 pub struct FmtArgs {
-    /// One or more .zen/.star files or directories containing such files to format.
-    /// When omitted, all .zen/.star files in the current directory tree are formatted.
+    /// One or more .zen files or directories containing .zen files to format.
+    /// When omitted, all .zen files in the current directory tree are formatted.
     #[arg(value_name = "PATHS", value_hint = clap::ValueHint::AnyPath)]
     pub paths: Vec<PathBuf>,
 
@@ -30,25 +30,25 @@ pub struct FmtArgs {
     pub hidden: bool,
 }
 
-/// Format a single file using buildifier
-fn format_file(buildifier: &Buildifier, file_path: &Path, args: &FmtArgs) -> Result<bool> {
+/// Format a single file using ruff formatter
+fn format_file(formatter: &RuffFormatter, file_path: &Path, args: &FmtArgs) -> Result<bool> {
     debug!("Formatting file: {}", file_path.display());
 
     if args.check {
-        buildifier.check_file(file_path)
+        formatter.check_file(file_path)
     } else if args.diff {
-        let diff = buildifier.diff_file(file_path)?;
+        let diff = formatter.diff_file(file_path)?;
         if !diff.is_empty() {
             print!("{diff}");
         }
         Ok(true)
     } else {
-        buildifier.format_file(file_path)?;
+        formatter.format_file(file_path)?;
         Ok(true)
     }
 }
 
-/// Recursively collect .zen and .star files from a directory
+/// Recursively collect .zen files from a directory
 fn collect_files_recursive(dir: &Path, files: &mut HashSet<PathBuf>, hidden: bool) -> Result<()> {
     for entry in fs::read_dir(dir)?.flatten() {
         let path = entry.path();
@@ -71,7 +71,7 @@ fn collect_files_recursive(dir: &Path, files: &mut HashSet<PathBuf>, hidden: boo
     Ok(())
 }
 
-/// Collect .zen and .star files from the provided paths
+/// Collect .zen files from the provided paths
 pub fn collect_files(paths: &[PathBuf], hidden: bool) -> Result<Vec<PathBuf>> {
     let mut unique: HashSet<PathBuf> = HashSet::new();
 
@@ -107,16 +107,11 @@ pub fn collect_files(paths: &[PathBuf], hidden: bool) -> Result<Vec<PathBuf>> {
 }
 
 pub fn execute(args: FmtArgs) -> Result<()> {
-    // Create a buildifier instance
-    let buildifier = Buildifier::new().context("Failed to initialize bundled buildifier")?;
+    // Create a ruff formatter instance
+    let formatter = RuffFormatter::default();
 
     // Print version info in debug mode
-    debug!(
-        "Using {}",
-        buildifier
-            .version()
-            .unwrap_or_else(|_| "buildifier".to_string())
-    );
+    debug!("Using ruff formatter");
 
     // Determine which files to format
     let starlark_paths = collect_files(&args.paths, args.hidden)?;
@@ -124,7 +119,7 @@ pub fn execute(args: FmtArgs) -> Result<()> {
     if starlark_paths.is_empty() {
         let cwd = std::env::current_dir()?;
         anyhow::bail!(
-            "No .zen or .star files found in {}",
+            "No .zen files found in {}",
             cwd.canonicalize().unwrap_or(cwd).display()
         );
     }
@@ -145,18 +140,12 @@ pub fn execute(args: FmtArgs) -> Result<()> {
             Spinner::builder(format!("{file_name}: Formatting")).start()
         };
 
-        match format_file(&buildifier, &file_path, &args) {
+        match format_file(&formatter, &file_path, &args) {
             Ok(is_formatted) => {
                 spinner.finish();
 
                 if args.check {
                     if is_formatted {
-                        println!(
-                            "{} {}",
-                            pcb_ui::icons::success(),
-                            file_name.with_style(Style::Green).bold()
-                        );
-                    } else {
                         println!(
                             "{} {} (needs formatting)",
                             pcb_ui::icons::warning(),
@@ -164,6 +153,12 @@ pub fn execute(args: FmtArgs) -> Result<()> {
                         );
                         all_formatted = false;
                         files_needing_format.push(file_path.clone());
+                    } else {
+                        println!(
+                            "{} {}",
+                            pcb_ui::icons::success(),
+                            file_name.with_style(Style::Green).bold()
+                        );
                     }
                 } else {
                     // For both diff mode and regular format mode, show success