--- a/crates/pcb/src/layout.rs
+++ b/crates/pcb/src/layout.rs
@@ -25,7 +25,7 @@ pub struct LayoutArgs {
     #[arg(value_name = "PATHS", value_hint = clap::ValueHint::AnyPath)]
     pub paths: Vec<PathBuf>,
 
-    /// Recursively traverse directories to find .zen/.star files
+    /// Recursively traverse directories to find .zen files
     #[arg(short = 'r', long = "recursive", default_value_t = false)]
     pub recursive: bool,