buildifier doesn't support f-strings, which we use in a bunch of places. We also run fmt after `pcb upgrade`, and it's a bit annoying that it fails due to a buildifier error.

it's also >10x faster than buildifier.
old:
```
> hyperfine "pcb fmt **/*.zen"
Benchmark 1: pcb fmt **/*.zen
  Time (mean ± σ):      59.5 ms ±   0.3 ms    [User: 33.0 ms, System: 22.8 ms]
  Range (min … max):    58.8 ms …  60.1 ms    47 runs
```

new:
```
hyperfine "./target/release/pcb fmt ../registry/**/*.zen"
Benchmark 1: ./target/release/pcb fmt ../registry/**/*.zen
  Time (mean ± σ):       5.5 ms ±   0.4 ms    [User: 2.5 ms, System: 2.6 ms]
  Range (min … max):     5.1 ms …   7.7 ms    347 runs
```

diff on registry w/ ruff: https://github.com/diodeinc/registry/compare/ruff-fmt?expand=1